remove need for authentification
remove ability to specific a report template or findings layout. This was done since the script creates a ptrac file that should be imported to add findings to an existing report. Since there is no report info in the Acunetix360 file other than a report name, it makes more sense to import it as findings instead of importing the ptrac as a report.